+++ src/main/kotlin/com/gabinote/coffeenote/field/domain/fieldType/ShortTextField.kt
@@ -1,0 +1,10 @@
+package com.gabinote.coffeenote.field.domain.fieldType
+
+object ShortTextField : TextField() {
+    override val key: String
+        get() = "SHORT_TEXT"
+
+    override val maxLength: Int = 100
+    override val messageTypeName: String = "Short Text"
+
+}
