+++ src/main/kotlin/com/gabinote/coffeenote/field/domain/fieldType/TextField.kt
@@ -3,10 +3,9 @@ package com.gabinote.coffeenote.field.domain.fieldType
 import com.gabinote.coffeenote.common.util.collection.CollectionHelper.firstOrEmptyString
 import com.gabinote.coffeenote.field.domain.attribute.Attribute
 
-object TextField : FieldType() {
-    override val key: String
-        get() = "TEXT"
-
+abstract class TextField : FieldType() {
+    open val maxLength: Int = 100
+    open val messageTypeName: String = "Text"
     override val fieldTypeAttributeKeys: Set<FieldTypeAttributeKey> = setOf()
 
     override fun validationValues(values: Set<String>, attributes: Set<Attribute>): List<FieldTypeValidationResult> {
@@ -15,17 +14,18 @@ object TextField : FieldType() {
             results.add(
                 FieldTypeValidationResult(
                     valid = false,
-                    message = "Text field can has only 1 value"
+                    message = "$messageTypeName field can has only 1 value"
                 )
             )
         }
 
         val value = values.firstOrEmptyString()
-        if (value.length > 100) {
+
+        if (value.length > maxLength) {
             results.add(
                 FieldTypeValidationResult(
                     valid = false,
-                    message = "Text field value cannot exceed 100 characters"
+                    message = "$messageTypeName field value cannot exceed $maxLength characters"
                 )
             )
         }
@@ -36,5 +36,6 @@ object TextField : FieldType() {
 
         return results
     }
+
 
 }
