+++ src/test/kotlin/com/gabinote/coffeenote/field/web/controller/FieldAdminApiControllerTest.kt
@@ -448,7 +448,7 @@ class FieldAdminApiControllerTest : WebMvcTestTemplate() {
                     val validDto = FieldCreateDefaultReqControllerDto(
                         name = "50자를 넘지 않는 이름",
                         icon = "50자를 넘지 않는 아이콘",
-                        type = "TEXT",
+                        type = "SHORT_TEXT",
                         attributes = setOf(
                             AttributeCreateReqControllerDto(
                                 key = "50자 넘지 않는 키",
